backend/requirements.txt
Django>=4.2,<5.0
django-cors-headers>=4.3
djangorestframework>=3.14
psycopg2-binary>=2.9
python-decouple>=3.8
gunicorn>=21.0
